+++ SurfaceTopography/Container/IO/CE.py
@@ -38,8 +38,9 @@ from .Reader import ContainerReaderBase
 
 
 class _ReadTopography(object):
-    def __init__(self, reader, **kwargs):
+    def __init__(self, reader, ignore_filters=False, **kwargs):
         self._reader = reader
+        self._ignore_filters = ignore_filters
         self._kwargs = kwargs
 
     def __call__(self):
@@ -57,9 +58,11 @@ class _ReadTopography(object):
         # 'squeezed' prefix to the data file key
         datafile_key = self._kwargs['datafile_key']
         topo_meta = self._kwargs['topo_meta']
-        if not datafile_key.startswith('squeezed'):
+        if not self._ignore_filters and not datafile_key.startswith('squeezed'):
             if 'height_scale' in topo_meta:
                 t = t.scale(topo_meta['height_scale'])
+            if 'fill_undefined_data_mode' in topo_meta:
+                t = t.interpolate_undefined_data(topo_meta['fill_undefined_data_mode'])
             if 'detrend_mode' in topo_meta:
                 t = t.detrend(topo_meta['detrend_mode'])
 
@@ -85,7 +88,7 @@ class CEReader(ContainerReaderBase):
     This reader imports digital surface twin containers from https://contact.engineering/.
     '''
 
-    def __init__(self, fn, datafile_keys=['original', 'squeezed-netcdf']):
+    def __init__(self, fn, datafile_keys=['original', 'squeezed-netcdf'], ignore_filters=False):
         """
         Read all surfaces in a contact.engineering container file and associated
         metadata. The container is a ZIP file with raw data files and a YAML file
@@ -101,6 +104,9 @@ class CEReader(ContainerReaderBase):
             starts with 'squeezed', the pipeline is not constructed from
             the metadata.
             (Default: ['original', 'squeezed-netcdf'])
+        ignore_filters : bool, optional
+            If True, the filter pipeline is not (re-)constructed from the metadata.
+            (Default: False)
 
         Returns
         -------
@@ -181,6 +187,7 @@ class CEReader(ContainerReaderBase):
 
                 readers += [
                     _ReadTopography(reader,
+                                    ignore_filters=ignore_filters,
                                     physical_sizes=physical_sizes,
                                     periodic=periodic,
                                     unit=unit,